Reorganize the files to their respective folders - Fix build path errors and clean qmake build file (.pro) - One QT module per line - Drop support for QT4 - Use nullptr instead of Q_NULLPTR; - Use override instead of Q_DECL_OVERRIDE; - Fix cannot find gpuonfire.svg
