+++ src/base/base.pri
@@ -1,0 +1,6 @@
+
+SOURCES += src/base/main.cpp \
+           src/base/nvidia/nvidia.cpp
+
+HEADERS += src/base/nvidia/nvidia.h \
+           src/base/nvidia/nvml.h
