+++ src/gui/gui.pri
@@ -1,0 +1,18 @@
+
+UI_DIR = src/gui/settings
+
+SOURCES += src/gui/mainwindow.cpp \
+           src/gui/settings/editprofile.cpp \
+           src/gui/settings/newprofile.cpp \
+           src/gui/qcustomplot.cpp \
+           src/gui/plotwidget.cpp
+
+HEADERS += src/gui/mainwindow.h \
+           src/gui/settings/editprofile.h \
+           src/gui/settings/newprofile.h \
+           src/gui/qcustomplot.h \
+           src/gui/plotwidget.h
+
+FORMS +=   src/gui/mainwindow.ui \
+           src/gui/settings/editprofile.ui \
+           src/gui/settings/newprofile.ui
