+++ src/gui/plotwidget.cpp
