+++ src/gui/plotwidget.h
