+++ src/gui/settings/editprofile.ui
@@ -60,7 +60,7 @@
   <customwidget>
    <class>QCustomPlot</class>
    <extends>QWidget</extends>
-   <header>qcustomplot.h</header>
+   <header>src/gui/qcustomplot.h</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
