+++ src/icons/minusicon.png
